requirements.txt
asgiref==3.6.0
Django==4.2
django-phonenumber-field==7.0.2
psycopg2-binary==2.9.6
sqlparse==0.4.3
djangorestframework==3.14.0
Markdown==3.4.3
django-filter==23.2
